change(android/engine): Rename to DownloadFileUtils
